flag to deactivate link creation since it sometimes causes trouble

+++ scripts/ini_KITHiggsToTauTauAnalysis.sh
@@ -20,16 +20,22 @@ scram setup tauspinner
 cp $KITHIGGSTOTAUTAUPATH/data/fastbdt.xml $CMSSW_BASE/config/toolbox/$SCRAM_ARCH/tools/selected/fastbdt.xml
 scram setup fastbdt
 
-if [ command -v symlinks > /dev/null 2>&1 ]; then
-	symlinks -c ${CMSSW_BASE}/external/${SCRAM_ARCH}/lib/ > /dev/null 2>&1
+if [ "$1" != "no-links" ]
+then
+	if [ command -v symlinks > /dev/null 2>&1 ]; then
+		symlinks -c ${CMSSW_BASE}/external/${SCRAM_ARCH}/lib/ > /dev/null 2>&1
+	else
+		for FILE in ${CMSSW_BASE}/external/${SCRAM_ARCH}/lib/*; do
+			if [ -L $FILE ]; then
+				LINKPATH=$(readlink $FILE)
+				rm -f $FILE
+				ln -s ${LINKPATH/$CMSSW_BASE/..\/..\/..} $FILE
+			fi
+		done
+	fi
 else
-	for FILE in ${CMSSW_BASE}/external/${SCRAM_ARCH}/lib/*; do
-		if [ -L $FILE ]; then
-			LINKPATH=$(readlink $FILE)
-			rm -f $FILE
-			ln -s ${LINKPATH/$CMSSW_BASE/..\/..\/..} $FILE
-		fi
-	done
+	echo "Link creation deactivated"
+
 fi
 # FastBDT
 # export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:${CMSSW_BASE}/src/FastBDT/
